+++ app/src/main/java/com/critt/trandroidlator/ui/AuthViewModel.kt
@@ -6,7 +6,7 @@ import com.google.firebase.auth.FirebaseUser
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 import androidx.lifecycle.viewModelScope
-import com.critt.trandroidlator.util.getIdTokenSafely
+import com.critt.trandroidlator.getIdTokenSafely
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.flow.MutableStateFlow
