+++ app/src/main/java/com/critt/trandroidlator/ui/components/DropdownSelector.kt
@@ -5,7 +5,7 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.material3.*
 import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
-import com.critt.trandroidlator.data.LanguageData
+import com.critt.trandroidlator.domain.LanguageData
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
