--- app/src/main/java/com/critt/trandroidlator/util/Constants.kt
@@ -1,7 +1,0 @@
-package com.critt.trandroidlator.util
-
-object Constants {
-    // Shared Preferences
-    const val PREFS_NAME = "enc_prefs"
-    const val PREFS_KEY_FB_TOKEN = "fb_token"
-}
