+++ build.gradle.kts
@@ -5,4 +5,5 @@ plugins {
     kotlin("plugin.serialization") version "1.9.22" apply false
     id("com.google.dagger.hilt.android") version "2.51.1" apply false
     id("com.google.gms.google-services") version "4.4.2" apply false
+    id("com.android.library") version "8.2.0" apply false
 }
