+++ data/src/main/java/com/critt/trandroidlator/data/di/SocketsModule.kt
@@ -1,0 +1,18 @@
+package com.critt.trandroidlator.data.di
+
+import com.critt.trandroidlator.data.SessionManager
+import com.critt.trandroidlator.data.TranslationSource
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+object SocketsModule {
+    @Provides
+    @Singleton
+    fun provideSocketsService(sessionManager: SessionManager): TranslationSource =
+        TranslationSource(sessionManager)
+}
